add in prefix vs context mode
This adds the radio button for prefix memories vs context memories. Prefix memories do not add much to the context as they are in the bot prefix and only used during inference of the string. Context memories are added to the full context. The benefit of context memories is that they can add depth to the conversation as what was remembered previously can be used in the context. This release sets the baseline for finding the bug in the gen reply saving of ltm.