Update commandhandler.py
Fixed issue with FILE_LOAD command when given a directory rather than a single file causing a crash instead of working as expected.
The downside is that by doing it this way, there's no way to avoid sending a message to the bot.
This can cause it's message count to change, and create memories you may not want, due to the way the addon is structured.
It also doesn't send anything of use to the bot back through the command_output, only a count of ingested documents. This is required since even a few ingested documents could vastly go over context limits.